General bugfixing and code improvements
