+++ src/App.css
@@ -1,3 +1,19 @@
+html,
+body,
+#root {
+  height: 100%;
+}
+
 .App {
   text-align: center;
+  height: 100%;
+}
+
+.main {
+  margin-bottom: 120px;
+  height: 100%;
+}
+
+* {
+  box-sizing: border-box;
 }
