+++ src/components/Hero/Hero.jsx
@@ -1,4 +1,4 @@
-import './Hero.css';
+import './Hero.scss';
 
 export default function Hero() {
   return (
