+++ src/styles/Favorites.scss
